Function fast_text::skipgram

pub fn skipgram(args: &HashMap<&str, &str>)

Provides functionality for generating skipgrams.

Include argument names as HashMap keys and argument values as HashMap values, e.g.: "input" : "sample_text.txt" "output" : "sample"

Documentation from fastText:

The following arguments are mandatory: -input training file path -output output file path

The following arguments are optional: -verbose verbosity level [2]

The following arguments for the dictionary are optional: -minCount minimal number of word occurences [5] -minCountLabel minimal number of label occurences [0] -wordNgrams max length of word ngram [1] -bucket number of buckets [2000000] -minn min length of char ngram [3] -maxn max length of char ngram [6] -t sampling threshold [0.0001] -label labels prefix [label]

The following arguments for training are optional: -lr learning rate [0.05] -lrUpdateRate change the rate of updates for the learning rate [100] -dim size of word vectors [100] -ws size of the context window [5] -epoch number of epochs [5] -neg number of negatives sampled [5] -loss loss function {ns, hs, softmax} [ns] -thread number of threads [12] -pretrainedVectors pretrained word vectors for supervised learning [] -saveOutput whether output params should be saved [0]

The following arguments for quantization are optional: -cutoff number of words and ngrams to retain [0] -retrain finetune embeddings if a cutoff is applied [0] -qnorm quantizing the norm separately [0] -qout quantizing the classifier [0] -dsub size of each sub-vector [2]